add: Added a new userscript to compare products
This userscript is highly experimental and has only been tested using the latest version of Firefox.

It adds a checkbox to each product found when browsing products inside a category. It only works on pages that does product listings, not highlighted products.
When the user clicks two checkboxes an overlay is presented to the user with the two products data sheets presented side by side.

It only works on products that have been updated to have the data field and only products that make sense to have date (games for example does not).

Known issues:
The formatting of the table columns are not equal and can be visually unpleasing.